ScriptEd: Go-to-definition works on everything now

+++ LegendaryExplorer/LegendaryExplorerCore/UnrealScript/Parsing/TokenStream.cs
@@ -151,7 +151,7 @@ namespace LegendaryExplorerCore.UnrealScript.Parsing
             return Data.GetEnumerator();
         }
 
-        public void AddDefinitionLink(ASTNode node, int offset, ushort length)
+        public void AddDefinitionLink(ASTNode node, int offset, int length)
         {
             DefinitionLinks.Add((node, offset, length));
         }
